Input mirror
Now let's write a simple program - the Input mirror. Your program needs to do the following
Accept an integer input from the console and store it in the variable N
Accept a string input from the console and store it in the variable S
Output the integer and the string on the same line
Solve the problem in the IDE and then click on click on Submit to proceed.