Number mirror - Negative integer
Let us now solve some programming problems. Note that
In the IDE - the '#' - comments will give you hints about what you need to do
The Solution tab also has '#' - comments which give you helpful information
Task
Write a program in the IDE which does the following
Accepts the count of test cases - t - in the 1st line
You need to generate the following output - Change the sign of N.
That is, if the input is 4, output −4. If the input is −5, output 5.