Debug this code - Determine the Score
The code given in the IDE is incorrect - Try and debug this program!!!
There is a problem worth X points and it has 10 test cases.
It is known that X is a multiple of 10.
It is known that each test case is worth the same number of points.
You pass N test cases among them. Determine the score that you will get.